requirements.txt
docx2python>=3.5.0
python-docx>=1.2.0
lxml>=6.0.0
langgraph-api>=0.2.3
langgraph-sdk>=0.1.61
botbuilder-core>=4.15.0
botbuilder-schema>=4.15.0
aiohttp>=3.8.0
requests>=2.28.0
python-dotenv>=0.19.0